Simplify home inventory into clear owners with typed preflight.
Split the inventory god module, replace stringly overwrite matching with
PreflightOutcome, and reuse matching home archives for remote tips so v1
acceptance stays proven without speculative CLI surface.
