+++ src/git.rs
@@ -8,6 +8,49 @@ use tempfile::TempDir;
 use crate::error::Error;
 use crate::sources::RemoteSource;
 
+/// Resolve the remote default branch tip without a full clone.
+pub fn remote_head(remote: &RemoteSource) -> Result<String, Error> {
+    let output = Command::new("git")
+        .args(["ls-remote", "--quiet", &remote.url, "HEAD"])
+        .env("GIT_TERMINAL_PROMPT", "0")
+        .output()
+        .map_err(|e| {
+            if e.kind() == std::io::ErrorKind::NotFound {
+                Error::msg("Git is required for remote skill sources")
+            } else {
+                Error::msg(format!("git ls-remote: {e}"))
+            }
+        })?;
+    if !output.status.success() {
+        let stderr = String::from_utf8_lossy(&output.stderr);
+        let detail = stderr
+            .lines()
+            .last()
+            .unwrap_or("git ls-remote failed")
+            .trim();
+        return Err(Error::msg(format!(
+            "Could not resolve {}: {detail}",
+            remote.url
+        )));
+    }
+    let stdout = String::from_utf8_lossy(&output.stdout);
+    let revision = stdout
+        .lines()
+        .next()
+        .and_then(|line| line.split_whitespace().next())
+        .ok_or_else(|| Error::msg(format!("Could not resolve HEAD for {}", remote.url)))?
+        .to_string();
+    if revision.len() != 40 && revision.len() != 64
+        || !revision.chars().all(|c| c.is_ascii_hexdigit())
+    {
+        return Err(Error::msg(format!(
+            "Could not resolve HEAD for {}: got {revision}",
+            remote.url
+        )));
+    }
+    Ok(revision)
+}
+
 /// Clone `remote` into a temporary directory; return `(temp_keep_alive, repo_path, revision)`.
 pub fn checkout(remote: &RemoteSource) -> Result<(TempDir, PathBuf, String), Error> {
     let temp = TempDir::new().map_err(|e| Error::msg(format!("temp dir: {e}")))?;
